requirement.txt
langchain
langchain-core
python-dotenv
langchain-google-genai
google-generativeai
# Hugging Face Integration
langchain-huggingface
transformers
huggingface-hub
langchain-anthropic

sentence_transformers

langchain_huggingface

faiss-cpu
chromadb
langchain_chroma

beautifulsoup4

streamlit


langchain_groq

fastapi

uvicorn

langserve

sse_starlette
langchain_community
pydantic

python-multipart

jinja2